Start implementing router protocol
Add initial wire protocol
Add a random cookie to the endpoint
Send cookie in connect request. Lower cookie wins. We do not yet
handle the case where the cookies are tied. Perhaps lower URI?

The wire protocol must not clash with the routed connection protocol
(still to be written). For now, the message type for both will be
in the first 8 bits. The router protocol will use the lower 4 bits and
the routed protocol will use the upper 4 bits. This will probably
change.

Routers may race when opening connections. We allow each peer to have
two (one active and one passive), but one will be rejected. Both will
never be fully connected at the same time.