Receive RIR msgs and store the routers in a tree
Add the topological structure which contains trees for the routers,
subnets, and routes.
Change BYE msgs to DELete msgs so that we can forward them. Eventually,
include the router id of the leaving server. Add a BYE flag to indicate
the router is a peer and will disconnect as well.