Stub out send completion handlers
Use the lower 3 bits of the send context to indicate what type of
context it is. Add flag for RCONN. Add GET_CTX macro.

For peer sends, we usually do not include a context and we set the
silent flag. If there is an error, we resend later anyway.

For E2E sends, we need to know the completion status. Store the rconn
for MSGs and the rma state for RMAs. On any error, tear down the
connection. We may want to have an exception for RNR and send that back
to the sender/initiator.